+++ carla_ros_bridge/src/carla_ros_bridge/traffic_lights_sensor.py
@@ -9,7 +9,8 @@
 a sensor that reports the state of all traffic lights
 """
 
-from carla_msgs.msg import CarlaTrafficLightStatusList
+from carla_msgs.msg import CarlaTrafficLightStatusList,\
+    CarlaTrafficLightInfoList
 from carla_ros_bridge.traffic import TrafficLight
 from carla_ros_bridge.pseudo_actor import PseudoActor
 
@@ -32,8 +33,10 @@ class TrafficLightsSensor(PseudoActor):
 
         super(TrafficLightsSensor, self).__init__(parent=parent,
                                                   communication=communication,
-                                                  prefix='traffic_lights')
+                                                  prefix="")
         self.actor_list = actor_list
+        self.traffic_light_status = CarlaTrafficLightStatusList()
+        self.traffic_light_actors = []
 
     def destroy(self):
         """
@@ -47,9 +50,23 @@ class TrafficLightsSensor(PseudoActor):
         """
         Get the state of all known traffic lights
         """
-        traffic_lights = CarlaTrafficLightStatusList()
+        traffic_light_status = CarlaTrafficLightStatusList()
+        traffic_light_actors = []
         for actor_id in self.actor_list:
             actor = self.actor_list[actor_id]
             if isinstance(actor, TrafficLight):
-                traffic_lights.traffic_lights.append(actor.get_status())
-        self.publish_message(self.get_topic_prefix(), traffic_lights)
+                traffic_light_actors.append(actor)
+                traffic_light_status.traffic_lights.append(actor.get_status())
+
+        if traffic_light_actors != self.traffic_light_actors:
+            self.traffic_light_actors = traffic_light_actors
+            traffic_light_info_list = CarlaTrafficLightInfoList()
+            for traffic_light in traffic_light_actors:
+                traffic_light_info_list.traffic_lights.append(traffic_light.get_info())
+            self.publish_message(self.get_topic_prefix() + "traffic_lights_info",
+                                 traffic_light_info_list, is_latched=True)
+
+        if traffic_light_status != self.traffic_light_status:
+            self.traffic_light_status = traffic_light_status
+            self.publish_message(self.get_topic_prefix() + "traffic_lights",
+                                 traffic_light_status, is_latched=True)