+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Latest changed
 
+*   Traffic Lights: Only publish to /carla/traffic_lights on change
+*   Traffic Lights: Publish /carla/traffic_lights_info, containing the location and the trigger volume
 ## CARLA-ROS-Bridge 0.9.8
 
 *   change Lidar range in meters